refactor: 💡 removed subfolders in sessionListItem component
