+++ src/components/sessionsListItem/_sessionsListItem.styles.scss
