+++ src/resources/scss/styles.scss
@@ -56,7 +56,7 @@
 @import '../../components/session/scss/session.styles';
 @import '../../components/session/scss/session.yellowTheme.styles';
 @import '../../components/sessionsList/scss/sessionsList.styles';
-@import '../../components/sessionsListItem/scss/sessionsListItem.styles';
+@import '../../components/sessionsListItem/sessionsListItem.styles';
 @import '../../components/typingIndicator/scss/typingIndicator.styles';
 @import '../../components/postcodeSuggestion/scss/postcodeSuggestion.styles';
 @import '../../components/login/scss/login.styles';
